refactor(frontend): update language dialogs
